dissolved extension methods in shapes_ext.dart split off the relevant normalization-only extension methods into normalize_shape.dart made the remainder class methods in shapes.dart
